Ticket #31: Coded fix to commands in Blackjack
I narrowed the problem down to Blackjack.get_bet. I was going to remove the ability to use commands there,
but that is the perfect time for the rpn or quit commands. Instead, I created a loop in get_bet that only
exits after a bet has been entered or a command ends the turn (returns False).