Issue #51: Fixed unchanged parameters on match functions.
The parameter list was changed when I start this, to be more consistent. I
forgot to go back and change the ones in monte_carlo_game.py, and I forgot to
change how they are called by Solitaire.match_check. Went back and fixed all
of that after finding the problem testing Canfield / variant = Selective.