Issue #55: Fixed match guessing in MultiSolitaire.
When I added matching to guessing, I did it in a way that assumed Solitaire.
It then broke when using MultiSolitaire when it got a list of cards instead of
a card. Switched it to use the do_ methods, which are overridden by
MultiSolitaire.

This was actually done in the crescent-dev branch which never got merged back
in. That suggests an error in my flow. Research. Fixed with new branch off
master and merge master back to the current -dev?