Issue #58: Redid Solitaire.default and .guess to better handle match.
This also involved writing a new guess_two method for guesses with two cards,
and redoing do_match and match_check to take two individual card parameters
(which is more consistent with do_build/build_check). Other changes were made
beyond just handling matches, such as reworking match_check to be a loop over
the two cards rather than duplicating the check code.

Note that the priorities are no longer as distinct for guess. Before it was a
strict sort/build/free/lane. But now build is replaced with build or match.
Build is a slight higher priority, but it checks both in each pile before
checking the next pile.