Issue #11: Modified LiarsDice bot and fixed number_word(0).
I changed ABBots' believable matrix. It refused to accept threes on first roll
and wouldn't accept that trip could become a trip.

The problem with utility.number_word was that it was looping while n, so if n
was 0, it never changed word. Added an if clause to fix that.