Issue #11: Trying to fix being unable to challenge in LiarsDice.
The problem was caused by not removing players. The challenge would go to the
next player, who wasn't in the game. I added code to find the next player by
going until there was a player with a score. However, now it is not doing the
player after the challenge correctly. I challenged, they lied, and then they
rolled again.