Issue #6: Coded one-suit and two-suit options for Spider.
This required adding one and two suit cards to cards.py, which worked nicely.
It also gave me a chance to test the sorting, which needed some fixes and
needs some more. At this point it doesn't trigger on alt moves, a significant
problem. I want to refactor the code that checks a move, and override
do_alternate to trigger it from there as well as do_build.

Then there is one more option, relaxed-turn (turn with empty piles).