Issue #146: Fixed storing sets in player.py.
The problem wasn't storing, exactly. The appending done by
store_shortcut assumes there is a blank line at the end of the file.
If that isn't there, the append won't work right. So I added a check
when the file is loaded to make sure there is a blank line at the
end of it.