Issue #230: Wrote code for full TrackingCard text with location.
There are problems. One, all of MultiSolitaire is not using the full text
versions. Two, it shouldn't be on the abbreviated text, because that is
messing up the code guess creates. Well, guess shouldn't be using that,
but still. Maybe the abbreviated versions should have their own way of
showing the location.